🧹Rename validation schemas folder
